# [SPARK-6951][core] Speed up parsing of event logs during listing.